Future: do not let then and andThen cast the continuation's futures #40589
Until now, there was an implicit cast of the resulting future of the
continuations, producing a warning. This cast was dangerous because it
could block the current thread whereas it was not desired.

Instead of unwrapping the future whenever possible, it was decided that
to let the users of then and andThen decided what results to unwrap, so
that to prevent losing information.

For this reason, in this fix we return the wrapped value, even if it
consists in an API change, breaking the compilation. All parts of libqi
that were broken are fixed in this commit.

As a side effect, the strands won't unwrap the futures anymore. It is also
more coherent. But for qi::bind to keep working a "strandedUnwrapped"
function had to be defined on actors, along with a "unwrappedSchedulerFor"
function on strands.

The need for qi::bind is for deprecated functions only, but it conflicts
with other parts of the code, like trackable.hxx, which need qi::Future
to implement qi::bind!

By putting further away the declarations requiring qi::bind, we allow
trackable to react properly to futures, and make more flexible this
include knot.

Change-Id: I9b732a15421c4508fd64d400a48840948970e791
Reviewed-on: http://gerrit.aldebaran.lan/86502
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>